modified report.md

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        long current;
+        int current;
         current = 2_000_000_000;
         int transfer;
         transfer = 500_000_000;
